+++ packages/uql-orm/src/index.ts
@@ -2,7 +2,6 @@ export * from './context/context.js';
 export * from './dialect/index.js';
 export * from './entity/index.js';
 export * from './namingStrategy/index.js';
-export * from './options.js';
 export * from './querier/index.js';
 export * from './type/index.js';
 export * from './util/index.js';